First step in trying to do machine learning.
Generated some job data in the format of "title salary", in (Orange
parlance) BASKET and then in SPARSE MATRIX format.

Current progress is learning of discrete data has no issues, but
continuous data results topped at about 65% even with 100000 sets of data.

Reasons unknown.

Code is not commented or optimized, has to be refactored and commented for
maintainability and readability for later stages.

TODO:
- Find out why continuous learning topped at 65%.
- Find out how to improve continous learning.